requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tornado==6.3.3
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
